requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
typer>=0.12.0
rich>=13.0.0
python-dotenv>=1.0.0
websockets>=13.0
